feat: Add i18n support to VaultPageContent
This commit adds internationalization (i18n) support to the `VaultPageContent` component. It imports the `useTranslation` hook from the `react-i18next` library and initializes the `t` function for translating text. The `match` function from the `match.ts` file is also imported. The `t` function is used to translate the text displayed in the button based on the `view` state. If the `view` is set to "balances", the selected vault's name is displayed. If the `view` is set to "vaults", the translated string "vaults" is displayed.