feat: Add QR code upload page and prompt button
This commit adds a new QR code upload page and a prompt button in the Vault page. The QR code upload page will allow users to upload QR codes for various purposes. The prompt button, represented by a camera icon, will be displayed at the bottom of the Vault page and will navigate users to the QR code upload page when clicked. This feature enhances the functionality of the Vault page by providing a convenient way for users to upload QR codes.