feat: Add styled-components utility for horizontal padding
This commit adds a new utility function, `horizontalPadding`, to the `styled-components` CSS library. The function takes a value (either a string or number) and generates CSS code for setting the left and right padding of an element. This utility can be used to easily apply consistent horizontal padding to elements in the UI, improving the overall layout and spacing.