feat: Add fill prop to Stack component
The Stack component in the frontend/src/lib/ui/layout/Stack.tsx file has been updated to include a new prop called "fill". When the "fill" prop is set to true, the Stack component will have a flex value of 1, allowing it to fill the available space within its parent container. This enhancement improves the flexibility and responsiveness of the Stack layout, enabling better utilization of available space.