feat: Add getBlockExplorerUrl utility function
This commit adds a new utility function called getBlockExplorerUrl to the chain/utils directory. This function is responsible for generating the block explorer URL for a given chain, entity, and value. It uses the match function from the lib/utils/match module to determine the appropriate URL based on the chain and entity.

The getBlockExplorerUrl function is used in the VaultAddressLink component to generate the block explorer URL for a vault address. When the user clicks on the "Block explorer" button in the VaultChainPage, the getBlockExplorerUrl function is called with the current chainId, the entity "address", and the vault address value. The generated URL is then opened in the browser using the BrowserOpenURL function from the wailsjs/runtime module.

This commit does not introduce any breaking changes and is backward-compatible with existing code.