feat: Add isNativeCoin utility function
This commit adds the utility function isNativeCoin to the chain/utils directory. The function checks if a given coin key is a native token by comparing it with the TokenSelectionAssets in the TokensStore. This function will be used in other parts of the codebase to determine if a coin is a native token.

Refactor VaultChainCoinItem component

This commit refactors the VaultChainCoinItem component in the vault/chain directory. The following changes were made:
- Replaced ChainEntityIcon with ChainCoinIcon from the chain/ui directory.
- Imported getChainEntityIconSrc from the chain/utils directory.
- Imported isNativeCoin from the chain/utils directory.
- Added CoinKey type to the component props.
- Updated the component to conditionally render the chain icon based on whether the coin is a native token or not.

Update VaultChainCoinPage component

This commit updates the VaultChainCoinPage component in the vault/chain/coin directory. The following changes were made:
- Added the coinKey object to the props of the CoinItem component.